feat: add modules exercise
